Added control over sequences to the ros interface

+++ rosserial_libs/ros_lib/shape_msgs/MeshTriangle.h
@@ -22,7 +22,7 @@ namespace shape_msgs
     virtual int serialize(unsigned char *outbuffer) const
     {
       int offset = 0;
-      for( uint32_t i = 0; i < 3; i++){
+      for( uint8_t i = 0; i < 3; i++){
       *(outbuffer + offset + 0) = (this->vertex_indices[i] >> (8 * 0)) & 0xFF;
       *(outbuffer + offset + 1) = (this->vertex_indices[i] >> (8 * 1)) & 0xFF;
       *(outbuffer + offset + 2) = (this->vertex_indices[i] >> (8 * 2)) & 0xFF;
@@ -35,7 +35,7 @@ namespace shape_msgs
     virtual int deserialize(unsigned char *inbuffer)
     {
       int offset = 0;
-      for( uint32_t i = 0; i < 3; i++){
+      for( uint8_t i = 0; i < 3; i++){
       this->vertex_indices[i] =  ((uint32_t) (*(inbuffer + offset)));
       this->vertex_indices[i] |= ((uint32_t) (*(inbuffer + offset + 1))) << (8 * 1);
       this->vertex_indices[i] |= ((uint32_t) (*(inbuffer + offset + 2))) << (8 * 2);
